Move script generation funcs to new module.
